misc refactoring, splitting scenarios
