+++ pom.xml
@@ -42,10 +42,6 @@
                 <groupId>org.apache.maven.plugins</groupId>
                 <artifactId>maven-compiler-plugin</artifactId>
                 <configuration>
-                    <includes>
-                        <include>src/main/**/*.java</include>
-                        <include>src/main/**/*.scala</include>
-                    </includes>
                     <source>${maven.compiler.release}</source>
                     <target>${maven.compiler.release}</target>
                 </configuration>
